Uploading Initial Files
Initial files include:
calendar.html - Has the base form for the getting the user input needed for holidayapi. Also has a div dedicated to where the actual calendar will ultimately go.
calendar.css - Quick minimal css for the form
calendar.js - A basic GET request to retrieve base info for the this month from holidayapi which is displayed as text on calendar.html in the div reserved for the calendar.