Minor CSS Update
Tried working with the start date type input but couldn't get it in time so I put things back to using datepicker and did a quick dirty css update to remove the "next" and "previous" month buttons from the date picker. Ideally this will only be the display calendar, but wanted to get it in in time.